pyproject.toml
[tool.poetry]
name = "simple-tgbot"
version = "0.2.0"
description = "Telegram bot for antispam and etc."
authors = ["doatta"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.1.1"
pyyaml = "^6.0.1"
gitpython = "^3.1.43"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
tgbot = 'simple_tgbot.src.bot:main'